pyproject.toml
[project]
name = "foothold"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "numpy>=2.4.4",
    "pyyaml>=6.0.3",
    "torch>=2.11.0",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu128"

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
